chore: add StringHash ctor

+++ src-server/ll/core/io/Output.cpp
@@ -5,6 +5,6 @@
 
 namespace ll::io {
 void defaultOutput(std::string_view sv) {
-    WriteConsoleA(GetStdHandle(STD_OUTPUT_HANDLE), sv.data(), sv.size(), nullptr, nullptr);
+    WriteConsoleA(GetStdHandle(STD_OUTPUT_HANDLE), sv.data(), (DWORD)sv.size(), nullptr, nullptr);
 }
 } // namespace ll::io
